+++ src/ILVisualizer.Application/Common/Interfaces/IILParserService.cs
@@ -1,4 +1,4 @@
-﻿using ILVisualizer.Domain.Models.IL;
+﻿using ILVisualizer.Domain.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
