+++ src/ILVisualizer.Bot/DependencyInjection/HostConfiguration.cs
@@ -3,6 +3,7 @@ using DSharpPlus;
 using Finite.Commands;
 using Finite.Commands.Parsing;
 using ILVisualizer.Application.Common;
+using ILVisualizer.Application.Common.Config;
 using ILVisualizer.Application.Common.Interfaces;
 using ILVisualizer.Application.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection;
