+++ src/ILVisualizer.Domain/Models/ILInstruction.cs
@@ -1,11 +1,11 @@
-﻿using ILVisualizer.Domain.Enums.IL;
+﻿using ILVisualizer.Domain.Enums;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ILVisualizer.Domain.Models.IL
+namespace ILVisualizer.Domain.Models
 {
     public struct ILInstruction
     {
