+++ tests/ILVisualizer.UnitTests/Parsing/ILParserServiceTests.cs
@@ -1,7 +1,7 @@
 ﻿using ILVisualizer.Application.Common.Exceptions.Parser;
 using ILVisualizer.Application.Common.Services;
-using ILVisualizer.Domain.Enums.IL;
-using ILVisualizer.Domain.Models.IL;
+using ILVisualizer.Domain.Enums;
+using ILVisualizer.Domain.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
